Messaging!
This adds support for sending messages to users. For moderation
purposes, admins can see anyone's conversation. This doesn't include the
live-updating bit as we're not entirely sure how the html should look
for this and I don't wanna clutter things with the extra bits for
live-updating before that's at least close to ready.