Added regions to users.
This way you'll be able to see that the item you want is being sold in
America and then you'll be paying import fees and extra shipping to get
it across the ocean. Also updated the readme because it was missing lots
of information for anyone who wants to poke at the project.